Restructure estar into a shared library
