+++ HEN_HOUSE/estar/estar_dataParser.cpp
@@ -1,23 +1,93 @@
-#include <iomanip>
 #include <iostream>
-#include <math.h>
-#include <string.h>
-#include <cctype>
+#include <fstream>
+#include <iomanip>
 #include <assert.h>
+#include "estar_dataParser.h"
+#include "estar_dataTables.h"
+
 using namespace std;
 
 /*
-    This module is intended to parse and interpret a single chemical formula of a
-    compound/element. The final product of this module is a parseformula object.
-    The comments inside the definition of structure parseformula gives more information on the output.
+    The purpose of this module is to read data from the elementData.h
+    file and then to properly structure the data for usage in other modules.
+    There are 100 sets of data in elementData.h (indexed by atomic number)
+    where each set of data corresponds to each element.
+    * For example: the third set of data in elementData.h is the data for Lithium
 */
 
-struct parseformula {
-    string str_arr[100]; // stores the formula of elements present in a compound/element into an array
-    int num_arr[100];    // for a compound/element this stores the number of atoms
-    // of the corresponding elements in str_arr into an array
-    int elem_types;      // number of different types of elements present
-};
+ElementOscillatorData parseData() {
+    ElementOscillatorData ds;
+    ds.numLevelsStandard = 113; // number of elements in the standard energy grid
+
+    double arr[14532];
+    int arr_len = 14532;
+    int i = 0;
+    double x;
+    while (i < arr_len) {
+        arr[i] = elementData[i];
+        i = i + 1;
+    }
+    //
+
+    //--//--//--//--//
+    // In the snippet below, we structure the data in arrays
+    i = 0;
+    int j = 0;
+    int temp_nmax[100];
+    int nc_temp[26];
+    double bd_temp[26];
+    double rlos_temp[113];
+    while (j<arr_len) {
+        temp_nmax[i] = arr[j];
+        int k = j+2; // the second element of each set of data is 113 and thus we ignore the second element
+        int a = 0;
+        /*
+            temp_nmax[i] is the first element of each set and this is an integer
+            length of nc_tamp array is temp_nmax[i]
+        */
+        while (k < j + 2 + temp_nmax[i]) {
+            nc_temp[a] = arr[k];
+            a = a + 1;
+            k = k + 1;
+        };
+        int m = k;
+        a = 0;
+        /*
+            Recall temp_nmax[i] is the first element of each set and this is an integer
+            length of bd_tamp array is temp_nmax[i]
+        */
+        while (m < k + temp_nmax[i]) {
+            bd_temp[a] = arr[m];
+            a = a + 1;
+            m = m + 1;
+        };
+        int n = m;
+        a = 0;
+        while (n < m + ds.numLevelsStandard) {
+            rlos_temp[a] = arr[n]; // the remaining 113 numbers in each set is stored in rlos_temp array
+            a = a + 1;
+            n = n + 1;
+        };
+        //--//--//--//--//
+
+        //--------------------
+        // In this snippet we input the data from nc_temp, bd_temp and rlos_temp
+        // into the structure
+        // Note that i is used to keep track of the set of data we are dealing with
+        // and the range of i is 0 <= i <= 99
+        ds.nmax[i] = temp_nmax[i];
+        for (int p = 0; p < temp_nmax[i]; p++) {
+            ds.nc[i][p] = nc_temp[p];
+        };
+        for (int p = 0; p < temp_nmax[i]; p++) {
+            ds.bd[i][p] = bd_temp[p];
+        };
+        //--------------------
+        i = i + 1;
+        j = n;
+    };
+    return ds;
+}
 
 /*
     The output of parse() below is a parseformula object. It takes input a formula string
@@ -96,5 +166,4 @@ parseformula parse(string str) {
     pf.elem_types = j;
     return pf;
 };
-
