+++ HEN_HOUSE/estar/estar_dataTables.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+#include <map>
+
+extern double er[113];
+extern double atb[100];
+extern double poth[100];
+extern double potgas[9];
+extern double potcon[9];
+extern std::map<std::string, int> per_table;
+
+extern double elementData[14532];
+
