--- HEN_HOUSE/estar/modules/parse.h
@@ -1,5 +1,0 @@
-#include "parse.cpp"
-
-struct parseformula;
-
-parseformula parse(string str);
