--- HEN_HOUSE/estar/modules/parseDataFile.h
@@ -1,5 +1,0 @@
-#include "parseDataFile.cpp"
-
-struct data;
-
-data parseData();
