--- HEN_HOUSE/estar/modules/routine/bspol.h
@@ -1,3 +1,0 @@
-#include "bspol.cpp"
-
-double fbspol();
