--- HEN_HOUSE/estar/modules/routine/formula.h
@@ -1,7 +1,0 @@
-#include "formula.cpp"
-
-struct formula_calc;
-
-int atom_num();
-
-formula_calc fcalc();
