--- HEN_HOUSE/estar/modules/routine/formulaCalculations.h
@@ -1,3 +1,0 @@
-#include "formula_calculation.cpp"
-
-formula_calc getDataFromFormulae(int knmat, double rho, string *elementArray, double *massFraction, float *numOfAtoms, int NEP);
