--- HEN_HOUSE/estar/modules/routine/mixformula.h
@@ -1,11 +1,0 @@
-#include "mixformula.cpp"
-
-struct formula_calc;
-
-struct mixtureData;
-
-mixtureData getData();
-
-mixtureData getEgsMediaData(string *elementArray, double *massFraction, int NEP);
-
-formula_calc mixtureCalculation(double rho, string *elementArray, double *massFraction, int NEP);
