--- HEN_HOUSE/estar/modules/routine/scof.h
@@ -1,5 +1,0 @@
-#include "scof.cpp"
-
-struct scof;
-
-scof fscof(int nmax, double x[1200], double f[1200]);
