--- HEN_HOUSE/estar/modules/solverHelpers/headerfile.h
@@ -1,7 +1,0 @@
-class HelperFunctions {
-public:
-    // cutoff is used to find the cutoff as discussed in section 4 of the report
-    // however in ESTAR a slight variation of cutoff is used (also discussed in section 4 of report)
-
-    double cutoff(double *, double *, double *, int);
-};
