+++ HEN_HOUSE/specs/all_common.spec
@@ -65,12 +65,6 @@ EGS_SOURCEDIR = $(HEN_HOUSE)src$(DSEP)
 # ESTAR sources
 #
 ESTAR_DIR = $(HEN_HOUSE)estar$(DSEP)
-ESTAR_MODULES = $(ESTAR_DIR)modules$(DSEP)
-ESTAR_ROUTINE = $(ESTAR_MODULES)routine$(DSEP)
-ESTAR_S_HELPERS = $(ESTAR_MODULES)solverHelpers$(DSEP)
-
-# ESTAR linking keyword
-F77_CPP_LINKER = -lstdc++ -std=c++11
 
 # Utilities coming with EGSnrc
 #
